+++ spec/helpers/assertions/expectWebsockets.ts
@@ -2,5 +2,5 @@ import expectFile from '../expectFile.js'
 
 export default async function expectWebsockets() {
   await expectFile('howyadoin/src/api/conf/websockets.ts')
-  await expectFile('howyadoin/src/api/utils/ws.ts')
+  await expectFile('howyadoin/src/api/utils/AppWs.ts')
 }